textutil: add color map facilities
Introduce SetColorMap()/GetColorMap() API's. With this the user may
change the termbox's output (i.e with Output256) mode and set a broader
color map.

In the future - if needed - we could introduce an API to switch modes
and automatically remap the colors as required. For now, this API's
are good enough.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>