Don't consider hidden children on ChildAt
When dealing with mouse events we use ChildAt() to find the controls
in the clicked area, but we may have hidden elements in the same position
in that case we want to dispatch events to the visible one. This patch
fixes this issue by filtering out the non visible elements.

Signed-off-by: Leandro Dorileo <leandro.maciel.dorileo@intel.com>