Allow for No Border on Window
Keeps current API as-is by defaulting to the new BorderAuto BorderStyle
which keeps the default borders on Windows and defaults on none on Frames.

Allows setting a Window Border to BorderNone for a frameless Window.

Need to allow setting the border default in the Window Manager to avoid
the border from being drawn and then removed.

Signed-off-by: Mark D Horn <mark.d.horn@intel.com>